+++ CHANGES.rst
@@ -65,6 +65,7 @@ esa.esasky
 - Added custom ADQL and TAP+ functionality. [#2078]
 
 - Enabled download of INTEGRAL data products. [#2105]
+- Added Solar System Object functionality. [#2106]
 
 esa.hubble
 ^^^^^^^^^^